%,HighligthMatchingPair 		-- added some bracket like pairs Q/ and Q?				-- turned off. #Z (imap to \mathrm)			-- removed. -> and <- (imaps)			-- imaps to \rightarrow and \leftarrow. Conceal					-- added to color scheme. TabCompletion				-- added greek letters, but not to completion pool but 						only if one stands right before, to not destroy 						if sb is trying to close a bracket or math environment 						for this I made changes in atplib#CloseLastEnvironment 						and atplib#CloseLastBracket functions. Some other fixes in various places, + changes on the web page.

+++ ftplugin/ATP_files/makelatex.py
@@ -261,7 +261,7 @@ def reload_xpdf():
 
 def copy_back_output(tmpdir):
     os.chdir(tmpdir)
-    if os.path.exists(file_cp):
+    if os.path.exists(file_cp) and os.path.exists(basename+output_ext):
         shutil.copy(basename+output_ext, texfile_dir)
     os.chdir(texfile_dir)
